requirements.txt
# Enterprise-Grade Facial Authentication System - Dependencies

# ============================================
# Core Web Framework
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================
# Deep Learning Frameworks
# ============================================
torch>=2.6.0
torchvision>=0.21.0
onnx==1.15.0
onnxruntime==1.16.3  # CPU-only runtime for Render
facenet-pytorch==2.5.3
insightface==0.7.3

# ============================================
# Computer Vision
# ============================================
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
albumentations==1.3.1
scikit-image==0.22.0
mtcnn==0.1.1
retinaface-pytorch==0.0.1

# ============================================
# Machine Learning
# ============================================
scikit-learn==1.3.2
scipy==1.11.4
numpy==1.24.3
pandas==2.1.3

# ============================================
# Database & ORM
# ============================================
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
supabase==2.3.4

# ============================================
# Security & Encryption
# ============================================
pyjwt==2.8.0
argon2-cffi==23.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
slowapi==0.1.9
redis==5.0.1
PyYAML==6.0.1

# ============================================
# Visualization
# ============================================
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# ============================================
# Monitoring & Logging
# ============================================
prometheus-client==0.19.0
python-json-logger==2.0.7

# ============================================
# Demo Application
# ============================================
streamlit==1.29.0
streamlit-webrtc==0.47.1
pyarrow==14.0.1  # Using a pre-built wheel version

# ============================================
# Development & Testing
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
ipykernel==6.27.1
jupyter==1.0.0
notebook==7.0.6

# ============================================
# Additional ML Tools
# ============================================
timm==0.9.12  # For EfficientNet and other models
grad-cam==1.4.8
onnx-simplifier==0.4.35
tensorboard==2.15.1


